Add method of building documentation in docker container
Excerpt from the documentation explaining why using a docker container
to build the documentaion is a good idea

  | [The docker method] is the recommended method, since this project is
  | very sensitive to which `python` version and `pip` package versions
  | are installed.  Indeed, `Sphinx` (the toolkit used for automatically
  | generating the code documentation) needs to be able to import the
  | `DynAIkonTrap` module (and all its dependencies). It can only do this
  | if all the correct versions are installed. Else conflicts are
  | possible. We have also experienced documentation build failures with
  | specific versions of `numpy`.
  |
  | Using the `python:3.7` container allows us to have a tightly
  | controlled environment in which we know the documentation will build
  | correctly.  Also adds corresponding documentatinon inside

It is also essentially the same method that the CI pipeline uses to
build the documentation: it spawns a python:3.7 docker container and
executes virtually the same commands. See `.gitlab-ci.yml` for
comparison.

This commit also adds corresponding documentatinon inside

  docs/source/dev-docs/build-docs.rst

and also contains a (pandoc automatically converted) copy of this at

  docs/README.md

for those working from the console.

Since `docs/README.md` is automatically converted from
`docs/source/dev-docs/build-docs.rst`, it should not be directly edited.
There is a warning banner in `docs/README.md` explaining this.